rest_root/courses/serializers.py
from rest_framework import serializers
from .models import Course


class CourseSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Course
        fields = ["code", "name", "progression", "plan"]
